// sdk/packages/core/src/kerberosec-core/types.ts
export interface KerberoSecCoreStartInput
	extends Omit<StartSessionInput, "config" | "localRuntime"> {
	config: KerberoSecCoreStartConfig;
	localRuntime?: LocalRuntimeStartOptions;
}